requirements.txt
# core
Django==3.2.8
djangorestframework==3.12.4

# database
mysqlclient==2.0.1
django-mysql==3.8.1

# bs4
beautifulsoup4==4.10.0
html5lib
lxml

#documentation (swagger)
drf-yasg==1.20.0

# authentication
PyJWT==2.3.0

# media files
Pillow==9.0.0

# email sending
sendgrid==6.9.3

# extra-fields (base64 for photo, etc)
drf-extra-fields

# .env
python-dotenv==0.19.2

requests==2.27.1

numpy
pandas